refactor: side 디렉토리 명 수정 (side->team)
