+++ src/main/java/janggi/domain/piece/Piece.java
@@ -3,7 +3,7 @@ package janggi.domain.piece;
 import janggi.domain.Board;
 import janggi.domain.MovePath;
 import janggi.domain.Position;
-import janggi.domain.side.TeamType;
+import janggi.domain.team.TeamType;
 import java.util.Optional;
 
 public interface Piece {
