+++ src/main/java/janggi/domain/piece/Po.java
@@ -4,7 +4,7 @@ import janggi.domain.Board;
 import janggi.domain.Delta;
 import janggi.domain.MovePath;
 import janggi.domain.Position;
-import janggi.domain.side.TeamType;
+import janggi.domain.team.TeamType;
 import java.util.List;
 import java.util.Optional;
