+++ src/main/java/janggi/view/OutputView.java
@@ -1,7 +1,7 @@
 package janggi.view;
 
 import janggi.domain.Position;
-import janggi.domain.side.TeamType;
+import janggi.domain.team.TeamType;
 import janggi.dto.BoardSpot;
 import janggi.dto.BoardSpots;
 import java.util.Map;
