+++ src/test/java/janggi/domain/team/ChuTest.java
@@ -1,4 +1,4 @@
-package janggi.domain.side;
+package janggi.domain.team;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
